% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutheme.R
\name{fluidEuTheme}
\alias{fluidEuTheme}
\title{Create a customized Shiny UI with additional CSS and HTML}
\usage{
fluidEuTheme(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{div}} function.}
}
\value{
A \code{shiny::fluidPage} object with custom styling.
}
\description{
This function extends a Shiny UI by adding custom CSS and HTML from the
\code{inst/www/} directory of the \code{eutheme} package.
}
